tests/wpt/web-platform-tests/websockets/binary/001.html

<!doctype html>
<title>WebSockets: Send/Receive blob, blob size less than network array buffer</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../constants.js?pipe=sub></script>
<meta name="variant" content="">
<meta name="variant" content="?wss">
<div id=log></div>
<script>
async_test(function(t){
  var ws = new WebSocket(SCHEME_DOMAIN_PORT + '/echo');
  var data = "";
  var datasize = 10;
  ws.onopen = t.step_func(function(e) {
   ws.binaryType = "blob";
   data = new ArrayBuffer(datasize);
   ws.send(data);
  })
  ws.onmessage = t.step_func(function(e) {
    assert_true(e.data instanceof Blob);
    assert_equals(e.data.size, datasize);
    t.done();
  })

});
</script>
